requirements.txt
flask==2.3.2
sqlalchemy==2.0.19
flask-sqlalchemy==3.0.5
slack-bolt==1.18.0
python-dotenv==1.0.0
requests==2.31.0
psycopg2-binary==2.9.7 